Improve the cleanup to goBack motor or not
The Cleanup method had a goBack attribute width default value to True,
then when the method is called (cleanup( )) the motor go back
to the previous position and restore the values of the velocity
and acceleration.
When the cleanup method is callen width False (cleanup(False))
the values of the motor are restored but the motor don't move
to the previous position.