Add a waiting-time mesh macro
Add a new standard macro based on a standard mesh macro.
This macro implements a hook mechanism which provides the possibility
to wait for a certain time before starting the acquisition at each mesh point.
This waiting time includes the elapsed time between the post and the pre
acquisition hooks..