Change macros qExafs and qExafsPos
Change the macros names between them. The new name of the old
qExafs is qExafsTime.

Prepare to use time mode with new PLC0

Implement macro comp

Implement macro comp to calculate and set the coefficients of the
experimental scale on the Mythen.

Implement macro comp

Implement macro comp to calculate and set the coeficients of the
experimental scale on the Mythen.

Implement the mythencomp

Macro mythencomp to extract on a spec file the compensation of the
raw data.